+++ NetherGate/src/com/elmakers/mine/bukkit/plugins/nether/PlayerList.java
@@ -1,0 +1,14 @@
+package com.elmakers.mine.bukkit.plugins.nether;
+
+import com.elmakers.mine.bukkit.plugins.nether.dao.TeleportingPlayer;
+import java.util.ArrayList;
+
+public class PlayerList extends ArrayList<TeleportingPlayer>
+{
+	public static final long serialVersionUID = 1;
+	
+	public PlayerList()
+	{
+		
+	}
+}